+++ tests/test_permissions.py
@@ -88,30 +88,30 @@ class TestPermissions(unittest.TestCase):
         ]
         format_perms = format_permissions(test_perms, PermissionType.ALLOWED)
         expect_names = [
-            Permission.READ.value,
-            str(PermissionSet(Permission.READ, Access.ALLOW, Scope.RECURSIVE)),
             Permission.READ.value + "-" + Scope.MATCH.value,
             str(PermissionSet(Permission.READ, Access.ALLOW, Scope.MATCH)),
+            Permission.READ.value,
+            str(PermissionSet(Permission.READ, Access.ALLOW, Scope.RECURSIVE)),
             # no implicit name for denied
-            str(PermissionSet(Permission.READ, Access.DENY, Scope.RECURSIVE)),
             str(PermissionSet(Permission.READ, Access.DENY, Scope.MATCH)),
-            Permission.WRITE.value,
-            str(PermissionSet(Permission.WRITE, Access.ALLOW, Scope.RECURSIVE)),
+            str(PermissionSet(Permission.READ, Access.DENY, Scope.RECURSIVE)),
             Permission.WRITE.value + "-" + Scope.MATCH.value,
             str(PermissionSet(Permission.WRITE, Access.ALLOW, Scope.MATCH)),
+            Permission.WRITE.value,
+            str(PermissionSet(Permission.WRITE, Access.ALLOW, Scope.RECURSIVE)),
             # no implicit name for denied
-            str(PermissionSet(Permission.WRITE, Access.DENY, Scope.RECURSIVE)),
             str(PermissionSet(Permission.WRITE, Access.DENY, Scope.MATCH)),
+            str(PermissionSet(Permission.WRITE, Access.DENY, Scope.RECURSIVE)),
         ]
         expect_perms = [
-            PermissionSet(Permission.READ, Access.ALLOW, Scope.RECURSIVE, PermissionType.ALLOWED).json(),
             PermissionSet(Permission.READ, Access.ALLOW, Scope.MATCH, PermissionType.ALLOWED).json(),
-            PermissionSet(Permission.READ, Access.DENY, Scope.RECURSIVE, PermissionType.ALLOWED).json(),
+            PermissionSet(Permission.READ, Access.ALLOW, Scope.RECURSIVE, PermissionType.ALLOWED).json(),
             PermissionSet(Permission.READ, Access.DENY, Scope.MATCH, PermissionType.ALLOWED).json(),
-            PermissionSet(Permission.WRITE, Access.ALLOW, Scope.RECURSIVE, PermissionType.ALLOWED).json(),
+            PermissionSet(Permission.READ, Access.DENY, Scope.RECURSIVE, PermissionType.ALLOWED).json(),
             PermissionSet(Permission.WRITE, Access.ALLOW, Scope.MATCH, PermissionType.ALLOWED).json(),
-            PermissionSet(Permission.WRITE, Access.DENY, Scope.RECURSIVE, PermissionType.ALLOWED).json(),
+            PermissionSet(Permission.WRITE, Access.ALLOW, Scope.RECURSIVE, PermissionType.ALLOWED).json(),
             PermissionSet(Permission.WRITE, Access.DENY, Scope.MATCH, PermissionType.ALLOWED).json(),
+            PermissionSet(Permission.WRITE, Access.DENY, Scope.RECURSIVE, PermissionType.ALLOWED).json(),
         ]
         utils.check_all_equal(format_perms["permission_names"], expect_names, any_order=False)
         utils.check_all_equal(format_perms["permissions"], expect_perms, any_order=False)
@@ -222,15 +222,15 @@ class TestPermissions(unittest.TestCase):
                            msg="Don't allow any object that makes the permission name not explicitly defined.")
 
     def test_compare_and_sort_operations(self):
-        perm_rar = PermissionSet(Permission.READ, Access.ALLOW, Scope.RECURSIVE)
         perm_ram = PermissionSet(Permission.READ, Access.ALLOW, Scope.MATCH)
-        perm_rdr = PermissionSet(Permission.READ, Access.DENY, Scope.RECURSIVE)
+        perm_rar = PermissionSet(Permission.READ, Access.ALLOW, Scope.RECURSIVE)
         perm_rdm = PermissionSet(Permission.READ, Access.DENY, Scope.MATCH)
-        perm_war = PermissionSet(Permission.WRITE, Access.ALLOW, Scope.RECURSIVE)
+        perm_rdr = PermissionSet(Permission.READ, Access.DENY, Scope.RECURSIVE)
         perm_wam = PermissionSet(Permission.WRITE, Access.ALLOW, Scope.MATCH)
-        perm_wdr = PermissionSet(Permission.WRITE, Access.DENY, Scope.RECURSIVE)
+        perm_war = PermissionSet(Permission.WRITE, Access.ALLOW, Scope.RECURSIVE)
         perm_wdm = PermissionSet(Permission.WRITE, Access.DENY, Scope.MATCH)
-        perm_sorted = [perm_rar, perm_ram, perm_rdr, perm_rdm, perm_war, perm_wam, perm_wdr, perm_wdm]
+        perm_wdr = PermissionSet(Permission.WRITE, Access.DENY, Scope.RECURSIVE)
+        perm_sorted = [perm_ram, perm_rar, perm_rdm, perm_rdr, perm_wam, perm_war, perm_wdm, perm_wdr]
         perm_random = [perm_rdr, perm_wam, perm_wdm, perm_rar, perm_war, perm_ram, perm_wdr, perm_rdm]
 
         utils.check_val_equal(perm_rar, Permission.READ, msg="Should be equal because of defaults and name conversion")
@@ -240,7 +240,7 @@ class TestPermissions(unittest.TestCase):
         utils.check_all_equal(list(sorted(perm_random)), perm_sorted, any_order=False)
 
         # cannot sort elements with other type representations although 'PermissionSet' can, because they don't know it
-        # (ie: str.__lt__ and dict.__ls__ could call with 'other' being a 'PermissionSet' depending on random order)
+        # (ie: str.__lt__ and dict.__ls__ could be called with 'other' being 'PermissionSet' depending on random order)
         # can still do the individual compares though if the reference object is 'PermissionSet' and other is anything
         utils.check_val_equal(perm_sorted[0] < perm_sorted[1], True)
         utils.check_val_equal(perm_sorted[0] < perm_sorted[1].json(), True)